Add files via upload
Digit recognition is the intelligence of computers to recognize digits written . But it becomes one of the most challenging tasks for machines as digits are not perfect and can be made with many different: flavours, size, thickness. Thus, as a solution to this problem, Digit recognition model comes into picture. 
Many machine learning techniques have been employed to solve the  digit recognition problem. This Project focuses on Neural Network (NN) approaches. 
Among the three famous NN approaches: deep neural network (DNN), deep belief network (DBN) and convolution neural network (CNN), the specialization of CNN as compared to other NN of being able to detect pattern is what makes it so useful for recognizing the digits.
 Our goal is to implement a CNN based handwritten digit recognition model that uses the image of a digit and recognizes the digit present in the image. The performance is tested on DATASET . The network was trained and tested on 1000 numeral samples. 
We carried out extensive experiments and achieved a recognition accuracy of 1.0 %.
It is a handwritten digit recognition , the ability of a computer to recognize the human handwritten digits from different sources like image, papers, touch screen etc. 
And classify them into 10 predefined classes (0-9). This is a set of 70,000 small image which are handwritten and these are 28x28 pixel image.
 It is a hard task for the machine because handwritten digits are not perfect and can be made with many different flavours. 
So, the handwriting digit recognition is the solution to this problem which uses the image of a digit and recognizes the digit present in the image.